refactor: remove printAsCsvOld
