--- src/record/export/printers/printAsCsvOld/__tests__/encloseInDoubleQuotes.test.ts
@@ -1,7 +1,0 @@
-import { encloseInDoubleQuotes } from "../encloseInDoubleQuotes";
-
-describe("encloseInDoubleQuotes", () => {
-  it("should enclose in double quotes correctly", () => {
-    expect(encloseInDoubleQuotes(`test`)).toBe(`"test"`);
-  });
-});
