--- src/record/export/printers/printAsCsvOld/__tests__/escapeDoubleQuotes.test.ts
@@ -1,7 +1,0 @@
-import { escapeDoubleQuotes } from "../escapeDoubleQuotes";
-
-describe("escapeDoubleQuotes", () => {
-  it("should escape double quotes correctly", () => {
-    expect(escapeDoubleQuotes(`"test"`)).toBe(`""test""`);
-  });
-});
