--- src/record/export/printers/printAsCsvOld/__tests__/hasSubtable.test.ts
@@ -1,14 +1,0 @@
-import { hasSubtable } from "../hasSubtable";
-import { FieldsJson } from "../../../../../kintone/types";
-
-const fieldsJson: FieldsJson = require("./fixtures/fields.json");
-const subtableFieldsJson: FieldsJson = require("./fixtures/subtable_fields.json");
-
-describe("hasSubTable", () => {
-  it("should be true when includes SUBTABLE in fields.json", () => {
-    expect(hasSubtable(subtableFieldsJson.properties)).toBe(true);
-  });
-  it("should be false when not includes SUBTABLE in fields.json", () => {
-    expect(hasSubtable(fieldsJson.properties)).toBe(false);
-  });
-});
