--- src/record/export/printers/printAsCsvOld/constants.ts
@@ -1,3 +1,0 @@
-export const LINE_BREAK = "\n";
-export const SEPARATOR = ",";
-export const PRIMARY_MARK = "*";
