--- src/record/export/printers/printAsCsvOld/encloseInDoubleQuotes.ts
@@ -1,3 +1,0 @@
-export const encloseInDoubleQuotes: (fieldValue: string) => string = (
-  fieldValue
-) => `"${fieldValue ? fieldValue : ""}"`;
