--- src/record/export/printers/printAsCsvOld/escapeDoubleQuotes.ts
@@ -1,2 +1,0 @@
-export const escapeDoubleQuotes = (fieldValue?: string | null) =>
-  fieldValue ? fieldValue.replace(/"/g, '""') : "";
