--- src/record/export/printers/printAsCsvOld/hasSubtable.ts
@@ -1,8 +1,0 @@
-import { FieldProperties } from "../../../../kintone/types";
-
-export const hasSubtable: (fieldProperties: FieldProperties) => boolean = (
-  fieldProperties
-) =>
-  Object.values(fieldProperties).some(
-    (fieldProperty) => fieldProperty.type === "SUBTABLE"
-  );
