+++ src/record/import/parsers/parseCsv/fieldValue.ts
@@ -1,6 +1,6 @@
-import { LINE_BREAK } from "../../../export/printers/printAsCsvOld/constants";
 import * as Fields from "../../types/field";
 import { FieldProperties } from "../../../../kintone/types";
+import { LINE_BREAK } from "../../../export/printers/printAsCsv/constants";
 
 export const convertFieldValue = ({
   type,
